Extract matchers to a separate class
